Moved the Hazelcast Starter constructor classes to own package
There will be more constructor classes in the future and also another
new package for org.mockito.stubbing.Answer classes. To keep the base
package clearly arranged, we should move the constructor classes to
their own package as well.
