+++ hazelcast/src/test/java/com/hazelcast/test/starter/constructor/GreaterLessPredicateConstructor.java
@@ -14,33 +14,31 @@
  * limitations under the License.
  */
 
-package com.hazelcast.test.starter;
+package com.hazelcast.test.starter.constructor;
 
 import java.lang.reflect.Constructor;
 
 import static com.hazelcast.test.starter.HazelcastProxyFactory.proxyArgumentsIfNeeded;
 import static com.hazelcast.test.starter.ReflectionUtils.getFieldValueReflectively;
 
-public class MapEventConstructor extends AbstractStarterObjectConstructor {
+public class GreaterLessPredicateConstructor extends AbstractStarterObjectConstructor {
 
-    public MapEventConstructor(Class<?> targetClass) {
+    public GreaterLessPredicateConstructor(Class<?> targetClass) {
         super(targetClass);
     }
 
     @Override
-    Object createNew0(Object delegate) throws Exception {
+    Object createNew0(Object delegate)
+            throws Exception {
         ClassLoader starterClassLoader = targetClass.getClassLoader();
-        Class<?> memberClass = starterClassLoader.loadClass("com.hazelcast.core.Member");
-        Constructor<?> constructor = targetClass.getConstructor(Object.class, memberClass, Integer.TYPE, Integer.TYPE);
+        Constructor<?> constructor = targetClass.getDeclaredConstructor(String.class,
+                Comparable.class, Boolean.TYPE, Boolean.TYPE);
 
-        Object source = getFieldValueReflectively(delegate, "source");
-        Object member = getFieldValueReflectively(delegate, "member");
-        Object entryEventType = getFieldValueReflectively(delegate, "entryEventType");
-        Integer eventTypeId = (Integer) entryEventType.getClass().getMethod("getType").invoke(entryEventType);
-        Object numberOfKeysAffected = getFieldValueReflectively(delegate, "numberOfEntriesAffected");
-
-        Object[] args = new Object[]{source, member, eventTypeId, numberOfKeysAffected};
-
+        Object attributeName = getFieldValueReflectively(delegate, "attributeName");
+        Object value = getFieldValueReflectively(delegate, "value");
+        Object equal = getFieldValueReflectively(delegate, "equal");
+        Object less = getFieldValueReflectively(delegate, "less");
+        Object[] args = new Object[]{attributeName, value, equal, less};
         Object[] proxiedArgs = proxyArgumentsIfNeeded(args, starterClassLoader);
 
         return constructor.newInstance(proxiedArgs);
