+++ hazelcast/src/test/java/com/hazelcast/test/starter/constructor/MapEventConstructor.java
@@ -14,27 +14,33 @@
  * limitations under the License.
  */
 
-package com.hazelcast.test.starter;
+package com.hazelcast.test.starter.constructor;
 
 import java.lang.reflect.Constructor;
 
 import static com.hazelcast.test.starter.HazelcastProxyFactory.proxyArgumentsIfNeeded;
 import static com.hazelcast.test.starter.ReflectionUtils.getFieldValueReflectively;
 
-public class LifecycleEventConstructor extends AbstractStarterObjectConstructor {
+public class MapEventConstructor extends AbstractStarterObjectConstructor {
 
-    public LifecycleEventConstructor(Class<?> targetClass) {
+    public MapEventConstructor(Class<?> targetClass) {
         super(targetClass);
     }
 
     @Override
     Object createNew0(Object delegate) throws Exception {
         ClassLoader starterClassLoader = targetClass.getClassLoader();
-        Class<?> stateClass = starterClassLoader.loadClass("com.hazelcast.core.LifecycleEvent$LifecycleState");
-        Constructor<?> constructor = targetClass.getDeclaredConstructor(stateClass);
+        Class<?> memberClass = starterClassLoader.loadClass("com.hazelcast.core.Member");
+        Constructor<?> constructor = targetClass.getConstructor(Object.class, memberClass, Integer.TYPE, Integer.TYPE);
 
-        Object state = getFieldValueReflectively(delegate, "state");
-        Object[] args = new Object[]{state};
+        Object source = getFieldValueReflectively(delegate, "source");
+        Object member = getFieldValueReflectively(delegate, "member");
+        Object entryEventType = getFieldValueReflectively(delegate, "entryEventType");
+        Integer eventTypeId = (Integer) entryEventType.getClass().getMethod("getType").invoke(entryEventType);
+        Object numberOfKeysAffected = getFieldValueReflectively(delegate, "numberOfEntriesAffected");
+
+        Object[] args = new Object[]{source, member, eventTypeId, numberOfKeysAffected};
+
         Object[] proxiedArgs = proxyArgumentsIfNeeded(args, starterClassLoader);
 
         return constructor.newInstance(proxiedArgs);
