+++ hazelcast/src/test/java/com/hazelcast/test/starter/constructor/test/DataAwareEntryEventConstructorTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.test.starter.test;
+package com.hazelcast.test.starter.constructor.test;
 
 import com.hazelcast.core.EntryEventType;
 import com.hazelcast.core.Member;
@@ -25,7 +25,7 @@ import com.hazelcast.spi.serialization.SerializationService;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
-import com.hazelcast.test.starter.DataAwareEntryEventConstructor;
+import com.hazelcast.test.starter.constructor.DataAwareEntryEventConstructor;
 import com.hazelcast.util.UuidUtil;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
