+++ hazelcast/src/test/java/com/hazelcast/test/starter/constructor/test/EnumConstructorTest.java
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package com.hazelcast.test.starter.test;
+package com.hazelcast.test.starter.constructor.test;
 
-import com.hazelcast.core.LifecycleEvent;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
-import com.hazelcast.test.starter.LifecycleEventConstructor;
+import com.hazelcast.test.starter.constructor.EnumConstructor;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -29,15 +28,18 @@ import static org.junit.Assert.assertEquals;
 
 @RunWith(HazelcastParallelClassRunner.class)
 @Category({QuickTest.class, ParallelTest.class})
-public class LifecycleEventConstructorTest {
+public class EnumConstructorTest {
 
     @Test
-    public void lifecycleEventCloneTest() {
-        LifecycleEvent lifecycleEvent = new LifecycleEvent(LifecycleEvent.LifecycleState.MERGING);
+    public void enumCloneTest() {
+        EnumConstructor constructor = new EnumConstructor(TestEnum.class);
 
-        LifecycleEventConstructor constructor = new LifecycleEventConstructor(LifecycleEvent.class);
-        LifecycleEvent clonedLifecycleEvent = (LifecycleEvent) constructor.createNew(lifecycleEvent);
+        assertEquals(TestEnum.FOO, constructor.createNew(TestEnum.FOO));
+        assertEquals(TestEnum.BAR, constructor.createNew(TestEnum.BAR));
+    }
 
-        assertEquals(lifecycleEvent.getState(), clonedLifecycleEvent.getState());
+    public enum TestEnum {
+        FOO,
+        BAR
     }
 }
